+++ beetsplug/web/static/beets.js
@@ -147,7 +147,7 @@ var BeetsRouter = Backbone.Router.extend({
     },
     itemQuery: function(query) {
         var queryURL = query.split(/\s+/).map(encodeURIComponent).join('/');
-        $.getJSON('/item/query/' + queryURL, function(data) {
+        $.getJSON('item/query/' + queryURL, function(data) {
             var models = _.map(
                 data['results'],
                 function(d) { return new Item(d); }
@@ -161,7 +161,7 @@ var router = new BeetsRouter();
 
 // Model.
 var Item = Backbone.Model.extend({
-    urlRoot: '/item'
+    urlRoot: 'item'
 });
 var Items = Backbone.Collection.extend({
     model: Item
@@ -264,7 +264,7 @@ var AppView = Backbone.View.extend({
         $('#extra-detail').empty().append(extraDetailView.render().el);
     },
     playItem: function(item) {
-        var url = '/item/' + item.get('id') + '/file';
+        var url = 'item/' + item.get('id') + '/file';
         $('#player audio').attr('src', url);
         $('#player audio').get(0).play();